[MH3-31350] - Added proper typecast to SYS_TMR_TickCountGet

+++ apps/wolfmqtt_azure_demo/firmware/src/config/pic32mz_ef_sk/system/sys_time_h2_adapter.c
@@ -57,7 +57,7 @@ uint64_t SYS_TMR_TickCountGetLong(void)
         gTickConv = SYS_TIME_MSToCount(1);
     }
 
-	return SYS_TIME_Counter64Get() / gTickConv;
+	return SYS_TIME_Counter64Get() / (uint64_t)gTickConv;
 }
 
 uint32_t SYS_TMR_TickCounterFrequencyGet ( void )
